feat: Add Drive diagnostics tool
Implement a diagnostic tool to check the status of the Google Drive connection and report on the presence of critical files and folders. This includes checking for the main 'Ouroboros' folder and essential data files. A new 'Bug' icon has also been added to the UI to trigger this feature.